+++ pet-store/src/main/java/pet/store/entity/PetStore.java
@@ -20,7 +20,7 @@ import lombok.ToString;
 public class PetStore {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int petStoreId;
+  private Long petStoreId;
   private String petStoreName;
   private String petStoreAddress;
   private String petStoreCity;
